pyproject.toml
[project]
name = "sattosat"
version = "0.1.0"
description = "Satellite orbit visualization and conjunction analysis tool"
requires-python = ">=3.10"
dependencies = [
    "sgp4>=2.22",
    "numpy>=1.24",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[tool.uv]
# Python scripts are in the python/ directory
package = false
